Remove a bad whitespace/todo rule (#17)
There are two lint checks that map to the category whitespace/todo.  One
of them checks that there is not more than one space between "//" and
"TODO".  For example, the following fails

    // Fruits:
    // - Apple: ...
    //   ...
    //   ...
    // - Pear: ...
    //   ...
    //   TODO...

Given the above legitimate case, this check has some false positives, so
get rid of it entirely.  There is still the other check mapping to
whitespace/todo that remains valid.