Add exception to readability/braces for concepts (#18)
Defining a concept as below can lead to readability/braces errors:
```
template<typename T>
concept TestConcept = requires(T value) {
   // ...
};
```
due to the semicolon at the end. This semicolon is required, and we use
`// NOLINT` at every place where this matches.

This adds an exception to the readability/braces for this case.